+++ rana_qgis_plugin/widgets/settings_dialog.py
@@ -1,3 +1,5 @@
+import os
+
 from qgis.PyQt.QtWidgets import (
     QDialog,
     QDialogButtonBox,
@@ -13,15 +15,18 @@ from qgis.PyQt.QtWidgets import (
 
 from rana_qgis_plugin.constant import PLUGIN_NAME
 from rana_qgis_plugin.utils import is_writable
+from rana_qgis_plugin.utils_api import get_frontend_settings
 from rana_qgis_plugin.utils_settings import (
     base_url,
     cognito_client_id,
     cognito_client_id_native,
     hcc_working_dir,
+    rana_cache_dir,
     set_base_url,
     set_cognito_client_id,
     set_cognito_client_id_native,
     set_hcc_working_dir,
+    set_rana_cache_dir,
 )
 
 
@@ -41,20 +46,6 @@ class SettingsDialog(QDialog):
         self.url_lineedit = QLineEdit(base_url(), auth_group)
         auth_group.layout().addWidget(self.url_lineedit, 0, 1)
 
-        auth_group.layout().addWidget(
-            QLabel("Cognito client ID (for login using identity provider)"), 1, 0
-        )
-        self.cognito_client_id_lineedit = QLineEdit(cognito_client_id(), auth_group)
-        auth_group.layout().addWidget(self.cognito_client_id_lineedit, 1, 1)
-
-        auth_group.layout().addWidget(
-            QLabel("Cognito client ID native (for login using username-password)"), 2, 0
-        )
-        self.cognito_client_id_native_lineedit = QLineEdit(
-            cognito_client_id_native(), auth_group
-        )
-        auth_group.layout().addWidget(self.cognito_client_id_native_lineedit, 2, 1)
-
         layout.addWidget(auth_group)
 
         sim_group = QGroupBox("Simulation", self)
@@ -62,11 +53,29 @@ class SettingsDialog(QDialog):
         sim_group.layout().addWidget(QLabel("Working directory"), 0, 0)
         self.working_dir_le = QLineEdit(hcc_working_dir(), sim_group)
         sim_group.layout().addWidget(self.working_dir_le, 0, 1)
-        browse_pb = QPushButton("Browse", sim_group)
-        sim_group.layout().addWidget(browse_pb, 0, 2)
-        browse_pb.clicked.connect(self.browse)
+        workdir_browse_pb = QPushButton("Browse", sim_group)
+        sim_group.layout().addWidget(workdir_browse_pb, 0, 2)
+        workdir_browse_pb.clicked.connect(
+            lambda: self.browse("Working Directory", self.working_dir_le)
+        )
 
         layout.addWidget(sim_group)
+
+        files_group = QGroupBox("File storage", self)
+        files_group.setLayout(QGridLayout())
+        files_group.layout().addWidget(QLabel("Cache directory"), 0, 0)
+        self.cache_dir_le = QLineEdit(rana_cache_dir(), files_group)
+        files_group.layout().addWidget(self.cache_dir_le, 0, 1)
+        cachedir_browse_pb = QPushButton("Browse", files_group)
+        files_group.layout().addWidget(cachedir_browse_pb, 0, 2)
+        cachedir_browse_pb.clicked.connect(
+            lambda: self.browse("Cache Directory", self.cache_dir_le)
+        )
+        if rana_cache_dir() is None:
+            base_dir = os.path.join(os.path.expanduser("~"), "Rana")
+            self.cache_dir_le.setText(base_dir)
+
+        layout.addWidget(files_group)
 
         buttonBox = QDialogButtonBox(
             QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel
@@ -81,32 +90,41 @@ class SettingsDialog(QDialog):
         return self._authenticationSettingsChanged
 
     def accept(self) -> None:
-        if self.cognito_client_id_lineedit.text() != cognito_client_id():
-            self._authenticationSettingsChanged = True
-            set_cognito_client_id(self.cognito_client_id_lineedit.text())
-
-        if self.cognito_client_id_native_lineedit.text() != cognito_client_id_native():
-            self._authenticationSettingsChanged = True
-            set_cognito_client_id_native(self.cognito_client_id_native_lineedit.text())
-
         if self.url_lineedit.text() != base_url():
             self._authenticationSettingsChanged = True
             set_base_url(self.url_lineedit.text())
 
+            frontend_settings = get_frontend_settings()
+            if (
+                frontend_settings
+                and frontend_settings["default_client_id"]
+                and frontend_settings["native_client_id"]
+            ):
+                set_cognito_client_id(frontend_settings["default_client_id"])
+                set_cognito_client_id_native(frontend_settings["native_client_id"])
+            else:
+                QMessageBox.critical(
+                    self,
+                    "Error",
+                    "Can't fetch settings from this backend URL. Please enter a different URL and try again.",
+                )
+                return
+
         set_hcc_working_dir(self.working_dir_le.text())
+        set_rana_cache_dir(self.cache_dir_le.text())
 
         return super().accept()
 
-    def browse(self):
-        work_dir = QFileDialog.getExistingDirectory(
-            self, "Select Working Directory", self.working_dir_le.text()
+    def browse(self, dir_human_name: str, dir_box_le: QLineEdit):
+        dir = QFileDialog.getExistingDirectory(
+            self, f"Select {dir_human_name}", dir_box_le.text()
         )
-        if work_dir:
-            if not is_writable(work_dir):
+        if dir:
+            if not is_writable(dir):
                 QMessageBox.warning(
                     self,
                     "Warning",
                     "Can't write to the selected location. Please select a folder to which you have write permission.",
                 )
                 return
-            self.working_dir_le.setText(work_dir)
+            dir_box_le.setText(dir)