+++ HISTORY.rst
@@ -11,6 +11,13 @@ History
 - Replace page specific refresh buttons with single refresh button
 - Refresh files browser, file view and revisions view every 60 seconds
 - Add rename functionality
+- Add support for selecting and uploading multiple files to Rana
+- Add option to create a new folder
+- Add button and dialog from M&S plugin to generate/upload new schematisation to 3Di and Rana
+- Allow user to set file cache directory
+- Rename tenant to organisation in user interfaces
+- Reactivate plugin when workdir is not set
+= Dynamically fetch client cognito ids from Rana API endpoint
 
 
 1.1.20 (2025-11-18)